Move shared code to common folder
