+++ browser/brave_vpn/dns/BUILD.gn
@@ -11,7 +11,7 @@ source_set("unit_tests") {
   if (enable_brave_vpn && is_win) {
     sources = [ "brave_vpn_dns_observer_service_unittest.cc" ]
     deps = [
-      "//brave/components/brave_vpn",
+      "//brave/components/brave_vpn/common",
       "//chrome/common:constants",
       "//chrome/test:test_support",
       "//net",
