+++ browser/brave_vpn/dns/brave_vpn_dns_observer_service_unittest.cc
@@ -12,9 +12,9 @@
 #include "base/test/scoped_feature_list.h"
 #include "brave/browser/brave_profile_prefs.h"
 #include "brave/browser/brave_vpn/dns/brave_vpn_dns_observer_factory.h"
-#include "brave/components/brave_vpn/brave_vpn_utils.h"
-#include "brave/components/brave_vpn/features.h"
-#include "brave/components/brave_vpn/pref_names.h"
+#include "brave/components/brave_vpn/common/brave_vpn_utils.h"
+#include "brave/components/brave_vpn/common/features.h"
+#include "brave/components/brave_vpn/common/pref_names.h"
 #include "chrome/browser/net/secure_dns_config.h"
 #include "chrome/browser/net/secure_dns_util.h"
 #include "chrome/browser/net/stub_resolver_config_reader.h"
