+++ browser/extensions/api/settings_private/brave_prefs_util.cc
@@ -58,7 +58,7 @@
 #endif
 
 #if BUILDFLAG(ENABLE_BRAVE_VPN)
-#include "brave/components/brave_vpn/pref_names.h"
+#include "brave/components/brave_vpn/common/pref_names.h"
 #endif
 
 namespace extensions {
