+++ browser/ui/brave_browser_command_controller_browsertest.cc
@@ -33,7 +33,7 @@
 #if BUILDFLAG(ENABLE_BRAVE_VPN)
 #include "brave/browser/brave_vpn/brave_vpn_service_factory.h"
 #include "brave/components/brave_vpn/brave_vpn_service.h"
-#include "brave/components/brave_vpn/features.h"
+#include "brave/components/brave_vpn/common/features.h"
 #endif
 
 class BraveBrowserCommandControllerTest : public InProcessBrowserTest {
