+++ components/brave_vpn/brave_vpn_os_connection_api.cc
@@ -13,11 +13,11 @@
 #include "base/json/json_reader.h"
 #include "base/power_monitor/power_monitor.h"
 #include "brave/components/brave_vpn/brave_vpn_api_request.h"
-#include "brave/components/brave_vpn/brave_vpn_constants.h"
-#include "brave/components/brave_vpn/brave_vpn_data_types.h"
+#include "brave/components/brave_vpn/common/brave_vpn_constants.h"
+#include "brave/components/brave_vpn/common/brave_vpn_data_types.h"
 #include "brave/components/brave_vpn/brave_vpn_service_helper.h"
-#include "brave/components/brave_vpn/brave_vpn_utils.h"
-#include "brave/components/brave_vpn/pref_names.h"
+#include "brave/components/brave_vpn/common/brave_vpn_utils.h"
+#include "brave/components/brave_vpn/common/pref_names.h"
 #include "brave/components/brave_vpn/vpn_response_parser.h"
 #include "components/prefs/pref_service.h"
 #include "third_party/abseil-cpp/absl/types/optional.h"
