+++ components/brave_vpn/brave_vpn_service_helper.cc
@@ -14,9 +14,9 @@
 #include "base/ranges/algorithm.h"
 #include "base/time/time.h"
 #include "base/values.h"
-#include "brave/components/brave_vpn/brave_vpn_constants.h"
-#include "brave/components/brave_vpn/brave_vpn_data_types.h"
-#include "brave/components/brave_vpn/pref_names.h"
+#include "brave/components/brave_vpn/common/brave_vpn_constants.h"
+#include "brave/components/brave_vpn/common/brave_vpn_data_types.h"
+#include "brave/components/brave_vpn/common/pref_names.h"
 #include "brave/components/skus/browser/skus_utils.h"
 #include "components/prefs/pref_service.h"
 
@@ -49,28 +49,6 @@ mojom::Region GetRegionFromValue(const base::Value::Dict& value) {
 }
 
 }  // namespace
-
-// On desktop, the environment is tied to SKUs because you would purchase it
-// from `account.brave.com` (or similar, based on env). The credentials for VPN
-// will always be in the same environment as the SKU environment.
-//
-// When the vendor receives a credential from us during auth, it also includes
-// the environment. The vendor then can do a lookup using Payment Service.
-std::string GetBraveVPNPaymentsEnv(const std::string& env) {
-  if (env == skus::kEnvProduction)
-    return "";
-  // Use same value.
-  if (env == skus::kEnvStaging || env == skus::kEnvDevelopment)
-    return env;
-
-  NOTREACHED();
-
-#if defined(OFFICIAL_BUILD)
-  return "";
-#else
-  return "development";
-#endif
-}
 
 bool ValidateCachedRegionData(const base::Value::List& region_value) {
   for (const auto& value : region_value) {
@@ -201,43 +179,11 @@ bool IsValidCredentialSummary(const base::Value& summary) {
   return active && remaining_credential_count > 0;
 }
 
-bool HasValidSubscriberCredential(PrefService* local_prefs) {
-  const base::Value::Dict& sub_cred_dict =
-      local_prefs->GetDict(prefs::kBraveVPNSubscriberCredential);
-  if (sub_cred_dict.empty())
-    return false;
-
-  const std::string* cred = sub_cred_dict.FindString(kSubscriberCredentialKey);
-  const base::Value* expiration_time_value =
-      sub_cred_dict.Find(kSubscriberCredentialExpirationKey);
-
-  if (!cred || !expiration_time_value)
-    return false;
-
-  if (cred->empty())
-    return false;
-
-  auto expiration_time = base::ValueToTime(expiration_time_value);
-  if (!expiration_time || expiration_time < base::Time::Now())
-    return false;
-
-  return true;
-}
 
 bool HasSubscriberCredential(PrefService* local_prefs) {
   const base::Value::Dict& sub_cred_dict =
       local_prefs->GetDict(prefs::kBraveVPNSubscriberCredential);
   return !sub_cred_dict.empty();
-}
-
-std::string GetSubscriberCredential(PrefService* local_prefs) {
-  if (!HasValidSubscriberCredential(local_prefs))
-    return "";
-  const base::Value::Dict& sub_cred_dict =
-      local_prefs->GetDict(prefs::kBraveVPNSubscriberCredential);
-  const std::string* cred = sub_cred_dict.FindString(kSubscriberCredentialKey);
-  DCHECK(cred);
-  return *cred;
 }
 
 absl::optional<base::Time> GetExpirationTime(PrefService* local_prefs) {
