+++ components/brave_vpn/common/features.h
@@ -1,10 +1,10 @@
 /* Copyright (c) 2021 The Brave Authors. All rights reserved.
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_BRAVE_VPN_FEATURES_H_
-#define BRAVE_COMPONENTS_BRAVE_VPN_FEATURES_H_
+#ifndef BRAVE_COMPONENTS_BRAVE_VPN_COMMON_FEATURES_H_
+#define BRAVE_COMPONENTS_BRAVE_VPN_COMMON_FEATURES_H_
 
 #include "base/feature_list.h"
 #include "build/build_config.h"
@@ -20,4 +20,4 @@ BASE_DECLARE_FEATURE(kBraveVPNDnsProtection);
 }  // namespace features
 }  // namespace brave_vpn
 
-#endif  // BRAVE_COMPONENTS_BRAVE_VPN_FEATURES_H_
+#endif  // BRAVE_COMPONENTS_BRAVE_VPN_COMMON_FEATURES_H_
