+++ components/brave_vpn/common/pref_names.h
@@ -1,10 +1,10 @@
 /* Copyright (c) 2021 The Brave Authors. All rights reserved.
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_BRAVE_VPN_PREF_NAMES_H_
-#define BRAVE_COMPONENTS_BRAVE_VPN_PREF_NAMES_H_
+#ifndef BRAVE_COMPONENTS_BRAVE_VPN_COMMON_PREF_NAMES_H_
+#define BRAVE_COMPONENTS_BRAVE_VPN_COMMON_PREF_NAMES_H_
 
 #include "build/build_config.h"
 
@@ -44,4 +44,4 @@ constexpr char kBraveVPNDaysInMonthUsed[] =
 
 }  // namespace brave_vpn
 
-#endif  // BRAVE_COMPONENTS_BRAVE_VPN_PREF_NAMES_H_
+#endif  // BRAVE_COMPONENTS_BRAVE_VPN_COMMON_PREF_NAMES_H_
